% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oho.R
\name{oho}
\alias{oho}
\title{Short description of what this does}
\usage{
oho(x, y)
}
\arguments{
\item{x}{The first thing}

\item{y}{The second thing.}
}
\value{
The thing that comes out of this function
}
\description{
Longer description of what this does. Approximately a paragraph.
}
\examples{
oho(1, 10)
}
